Fixed crash on TabController's mainController switching
Changed TabPool to singleton-pattern (it was not exact one actually).
Separated mainController switching and controller unregistration.
Moved controller unregistration to be called in destructor (that being not was actually the bug -- controller was not removing itself in case it was just orphaned, not closed) and changed it to use raw pointer, since shared ptrs are not valid in destructor.